WIP speedup TICKF 2
I renamed _pstakeMark in the previous message, just so GHC would require me to update every location in which that field was set.

I built everything in `ouroboros-network`, but not in `cardano-node` (I don't have the deps set correctly right now).

This commit undoes the rename. The long term fix (instead of simply squashing
this commit) would be to use PatternSynonyms or some such to ensure the pool
distr thunk is updated everytime the mark snapshot changes.
(One option is to instead add the thunk to each snapshot.)